:technologist: improved dx (examples, slower AnimatedMeshGradient speed, more package exports)
