+++ lib/mesh_gradient.dart
@@ -4,9 +4,11 @@ library mesh_gradient;
 export 'package:mesh_gradient/src/widgets/animated_mesh_gradient/animated_mesh_gradient.dart';
 export 'package:mesh_gradient/src/widgets/animated_mesh_gradient/animated_mesh_gradient_options.dart';
 export 'package:mesh_gradient/src/widgets/animated_mesh_gradient/animated_mesh_gradient_controller.dart';
+export 'package:mesh_gradient/src/widgets/animated_mesh_gradient/animated_mesh_gradient_painter.dart';
 
 /// Mesh Gradient
 export 'package:mesh_gradient/src/widgets/mesh_gradient/mesh_gradient.dart';
 export 'package:mesh_gradient/src/widgets/mesh_gradient/mesh_gradient_point.dart';
 export 'package:mesh_gradient/src/widgets/mesh_gradient/mesh_gradient_options.dart';
 export 'package:mesh_gradient/src/widgets/mesh_gradient/mesh_gradient_controller.dart';
+export 'package:mesh_gradient/src/widgets/mesh_gradient/mesh_gradient_painter.dart';
